linear

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,14 +11,14 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 	if (array == NULL || size == 0)
 		return (-1);
 
-	for( i = 0; i < size; i++)
+	for( i = 0; i < (int)size; i++)
 		if (array[i] == value)
 		{
-			printf("Value checked: array[%zu] = [%d]\n", i, array[i]);
+			printf("Value checked: array[%d] = [%d]\n", i, array[i]);
                         return (i);
 		}
 		else
